pyproject.toml
[tool.poetry]
name = "vox-kafka"
version = "0.2.3"
description = "Sinple kafka abstraction"
authors = ["Jhonatan Teixeira <jhonatan.teixeira@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "*"
django-health-check = {version = "*", optional = true}
Django = {version = ">=3.0,<3.1", optional = true}

[tool.poetry.dev-dependencies]
coverage = "*"
invoke = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
